requirements.txt
# Core app
gradio==5.49.1
transformers==4.57.1
accelerate==1.10.1
huggingface-hub==0.35.3
safetensors>=0.6.2
tokenizers==0.22.1

# OCR + imaging
easyocr==1.7.2
pillow==11.3.0
opencv-python-headless==4.12.0.88

# PDF parsing
pymupdf==1.26.5
pdfplumber==0.11.7

# Data / validation
pydantic>=2.0,<2.12
numpy>=1.24.0,<3.0.0
pandas==2.3.3